+++ examples/cli/migrations/db.js
@@ -3,8 +3,9 @@
 
 // $ npm install redis
 // $ redis-server
-
+var Promise = require('bluebird')
 var redis = require('redis')
+Promise.promisifyAll(redis.RedisClient.prototype)
 var db = redis.createClient()
 
 module.exports = db