+++ test/cli.js
@@ -38,14 +38,11 @@ describe('$ migrate', function () {
   describe('init', function () {
     beforeEach(mkdirp.bind(mkdirp, TMP_DIR))
 
-    it('should create a migrations directory', function (done) {
-      init([], function (err, out, code) {
-        assert(!err)
-        assert.equal(code, 0)
-        assert.doesNotThrow(() => {
-          fs.accessSync(path.join(TMP_DIR, 'migrations'))
-        })
-        done()
+    it('should create a migrations directory', async function () {
+      let result = await init([])
+      assert.equal(result.code, 0)
+      assert.doesNotThrow(() => {
+        fs.accessSync(path.join(TMP_DIR, 'migrations'))
       })
     })
   }) // end init
@@ -53,189 +50,144 @@ describe('$ migrate', function () {
   describe('create', function () {
     beforeEach(mkdirp.bind(mkdirp, TMP_DIR))
 
-    it('should create a fixture file', function (done) {
-      create(['test'], function (err, out, code) {
-        assert(!err)
-        assert.equal(code, 0)
-        var file = out.split(':')[1].trim()
-        var content = fs.readFileSync(file, {
-          encoding: 'utf8'
-        })
-        assert(content)
-        assert(content.indexOf('module.exports.up') !== -1)
-        assert(content.indexOf('module.exports.down') !== -1)
-        done()
+    it('should create a fixture file', async function () {
+      let result = await create(['test'])
+      assert.equal(result.code, 0)
+      var file = result.out.split(':')[1].trim()
+      var content = fs.readFileSync(file, {
+        encoding: 'utf8'
       })
+      assert(content)
+      assert(content.indexOf('module.exports.up') !== -1)
+      assert(content.indexOf('module.exports.down') !== -1)
     })
 
-    it('should respect the --date-format', function (done) {
+    it('should respect the --date-format', async function () {
       var name = 'test'
       var fmt = 'yyyy-mm-dd'
       var now = formatDate(new Date(), fmt)
 
-      create([name, '-d', fmt], function (err, out, code) {
-        assert(!err)
-        assert.equal(code, 0)
-        assert.doesNotThrow(() => {
-          fs.accessSync(path.join(TMP_DIR, 'migrations', now + '-' + name + '.js'))
-        })
-        done()
+      let result = await create([name, '-d', fmt])
+      assert.equal(result.code, 0)
+      assert.doesNotThrow(() => {
+        fs.accessSync(path.join(TMP_DIR, 'migrations', now + '-' + name + '.js'))
       })
     })
 
-    it('should respect the --extention', function (done) {
+    it('should respect the --extention', async function () {
       var name = 'test'
       var fmt = 'yyyy-mm-dd'
       var ext = '.mjs'
       var now = formatDate(new Date(), fmt)
 
-      create([name, '-d', fmt, '-e', ext], function (err, out, code) {
-        assert(!err)
-        assert.equal(code, 0)
-        assert.doesNotThrow(() => {
-          fs.accessSync(path.join(TMP_DIR, 'migrations', now + '-' + name + ext))
-        })
-        done()
+      let result = await create([name, '-d', fmt, '-e', ext])
+      assert.equal(result.code, 0)
+      assert.doesNotThrow(() => {
+        fs.accessSync(path.join(TMP_DIR, 'migrations', now + '-' + name + ext))
       })
     })
 
-    it('should use the --template-file flag', function (done) {
-      create(['test', '-t', path.join(__dirname, 'util', 'tmpl.js')], function (err, out, code) {
-        assert(!err)
-        assert.equal(code, 0, out)
-        assert(out.indexOf('create') !== -1)
-        var file = out.split(':')[1].trim()
-        var content = fs.readFileSync(file, {
-          encoding: 'utf8'
-        })
-        assert(content.indexOf('test') !== -1)
-        done()
+    it('should use the --template-file flag', async function () {
+      let result = await create(['test', '-t', path.join(__dirname, 'util', 'tmpl.js')])
+      assert.equal(result.code, 0, result.out)
+      assert(result.out.indexOf('create') !== -1)
+      var file = result.out.split(':')[1].trim()
+      var content = fs.readFileSync(file, {
+        encoding: 'utf8'
       })
+      assert(content.indexOf('test') !== -1)
     })
 
-    it('should fail with non-zero and a helpful message when template is unreadable', function (done) {
-      create(['test', '-t', 'fake'], function (err, out, code) {
-        assert(!err)
-        assert.equal(code, 1)
-        assert(out.indexOf('fake') !== -1)
-        done()
-      })
+    it('should fail with non-zero and a helpful message when template is unreadable', async function () {
+      let result = await create(['test', '-t', 'fake'])
+      assert.equal(result.code, 1)
+      assert(result.out.indexOf('fake') !== -1)
     })
   }) // end create
 
   describe('up', function () {
-    it('should run up on multiple migrations', function (done) {
-      up([], function (err, out, code) {
-        assert(!err)
-        assert.equal(code, 0)
-        db.load()
-        assert(out.indexOf('up') !== -1)
-        assert.equal(db.numbers.length, 2)
-        assert(db.numbers.indexOf(1) !== -1)
-        assert(db.numbers.indexOf(2) !== -1)
-        done()
-      })
+    it('should run up on multiple migrations', async function () {
+      let result = await up([])
+      assert.equal(result.code, 0)
+      db.load()
+      assert(result.out.indexOf('up') !== -1)
+      assert.equal(db.numbers.length, 2)
+      assert(db.numbers.indexOf(1) !== -1)
+      assert(db.numbers.indexOf(2) !== -1)
     })
 
-    it('should run up to a specified migration', function (done) {
-      up(['1-one.js'], function (err, out, code) {
-        assert(!err)
-        assert.equal(code, 0)
-        db.load()
-        assert(out.indexOf('up') !== -1)
-        assert.equal(db.numbers.length, 1)
-        assert(db.numbers.indexOf(1) !== -1)
-        assert(db.numbers.indexOf(2) === -1)
-        done()
-      })
+    it('should run up to a specified migration', async function () {
+      let result = await up(['1-one.js'])
+      assert.equal(result.code, 0)
+      db.load()
+      assert(result.out.indexOf('up') !== -1)
+      assert.equal(db.numbers.length, 1)
+      assert(db.numbers.indexOf(1) !== -1)
+      assert(db.numbers.indexOf(2) === -1)
     })
 
-    it('should run up multiple times', function (done) {
-      up([], function (err, out, code) {
-        assert(!err)
-        assert.equal(code, 0)
-        db.load()
-        assert(out.indexOf('up') !== -1)
-        up([], function (err, out) {
-          assert(!err)
-          assert(out.indexOf('up') === -1)
-          assert.equal(db.numbers.length, 2)
-          done()
-        })
-      })
+    it('should run up multiple times', async function () {
+      let result = await up([])
+      assert.equal(result.code, 0)
+      db.load()
+      assert(result.out.indexOf('up') !== -1)
+      result = await up([])
+      assert(result.out.indexOf('up') === -1)
+      assert.equal(db.numbers.length, 2)
     })
 
-    it('should run down when passed --clean', function (done) {
-      up([], function (err, out, code) {
-        assert(!err)
-        assert.equal(code, 0)
-        up(['--clean'], function (err, out) {
-          assert(!err)
-          db.load()
-          assert(out.indexOf('down') !== -1)
-          assert(out.indexOf('up') !== -1)
-          assert.equal(db.numbers.length, 2)
-          done()
-        })
-      })
+    it('should run down when passed --clean', async function () {
+      let result = await up([])
+      assert.equal(result.code, 0)
+      result = await up(['--clean'])
+      db.load()
+      assert(result.out.indexOf('down') !== -1)
+      assert(result.out.indexOf('up') !== -1)
+      assert.equal(db.numbers.length, 2)
     })
   }) // end up
 
   describe('down', function () {
-    beforeEach(function (done) {
-      up([], done)
+    beforeEach(async function () {
+      await up([])
     })
-    it('should run down on multiple migrations', function (done) {
-      down([], function (err, out, code) {
-        assert(!err)
-        assert.equal(code, 0)
-        db.load()
-        assert(out.indexOf('down') !== -1)
-        assert.equal(db.numbers.length, 0)
-        assert(db.numbers.indexOf(1) === -1)
-        assert(db.numbers.indexOf(2) === -1)
-        done()
-      })
+    it('should run down on multiple migrations', async function () {
+      let result = await down([])
+      assert.equal(result.code, 0)
+      db.load()
+      assert(result.out.indexOf('down') !== -1)
+      assert.equal(db.numbers.length, 0)
+      assert(db.numbers.indexOf(1) === -1)
+      assert(db.numbers.indexOf(2) === -1)
     })
 
-    it('should run down to a specified migration', function (done) {
-      down(['2-two.js'], function (err, out, code) {
-        assert(!err)
-        assert.equal(code, 0)
-        db.load()
-        assert(out.indexOf('down') !== -1)
-        assert.equal(db.numbers.length, 1)
-        assert(db.numbers.indexOf(1) !== -1)
-        assert(db.numbers.indexOf(2) === -1)
-        done()
-      })
+    it('should run down to a specified migration', async function () {
+      let result = await down(['2-two.js'])
+      assert.equal(result.code, 0)
+      db.load()
+      assert(result.out.indexOf('down') !== -1)
+      assert.equal(db.numbers.length, 1)
+      assert(db.numbers.indexOf(1) !== -1)
+      assert(db.numbers.indexOf(2) === -1)
     })
 
-    it('should run down multiple times', function (done) {
-      down([], function (err, out, code) {
-        assert(!err)
-        assert.equal(code, 0)
-        assert(out.indexOf('down') !== -1)
-        db.load()
-        down([], function (err, out) {
-          assert(!err)
-          assert(out.indexOf('down') === -1)
-          assert.equal(db.numbers.length, 0)
-          done()
-        })
-      })
+    it('should run down multiple times', async function () {
+      let result = await down([])
+      assert.equal(result.code, 0)
+      assert(result.out.indexOf('down') !== -1)
+      db.load()
+      result = await down([])
+      assert(result.out.indexOf('down') === -1)
+      assert.equal(db.numbers.length, 0)
     })
   }) // end down
 
   describe('list', function () {
-    it('should list available migrations', function (done) {
-      list([], function (err, out, code) {
-        assert(!err)
-        assert.equal(code, 0, out)
-        assert(out.indexOf('1-one.js') !== -1)
-        assert(out.indexOf('2-two.js') !== -1)
-        done()
-      })
+    it('should list available migrations', async function () {
+      let result = await list([])
+      assert.equal(result.code, 0, result.out)
+      assert(result.out.indexOf('1-one.js') !== -1)
+      assert(result.out.indexOf('2-two.js') !== -1)
     })
   }) // end init
 })