+++ test/fixtures/env/migrations/1-add-guy-ferrets.js
@@ -1,15 +1,13 @@
 var db = require('../../../util/db')
 
-exports.up = function (next) {
+exports.up = async function () {
   db[process.env.DB].push({ name: 'tobi' })
   db[process.env.DB].push({ name: 'loki' })
   db.persist()
-  next()
 }
 
-exports.down = function (next) {
+exports.down = async function () {
   db[process.env.DB].pop()
   db[process.env.DB].pop()
   db.persist()
-  next()
 }