+++ test/fixtures/issue-33/3-migration.js
@@ -1,12 +1,10 @@
 
 var db = require('../../util/db')
 
-exports.up = function (next) {
+exports.up = async function () {
   db.issue33.push('3-up')
-  next()
 }
 
-exports.down = function (next) {
+exports.down = async function () {
   db.issue33.push('3-down')
-  next()
 }