+++ test/fixtures/numbers/migrations/1-one.js
@@ -1,31 +1,24 @@
 var assert = require('assert')
 var db = require('../../../util/db')
 
-exports.up = function (next) {
+exports.up = async function () {
   db.load()
   db.numbers.push(1)
   db.persist()
-  next()
 }
 
-exports.down = function (next) {
+exports.down = async function () {
   db.load()
   db.numbers.pop()
   db.persist()
-  next()
 }
 
-exports.test = function (next) {
-  exports.up(function (err) {
-    if (err) return next(err)
-    exports.verify(function (err) {
-      if (err) return next(err)
-      exports.down(next)
-    })
-  })
+exports.test = async function () {
+  await exports.up()
+  exports.verify()
+  return exports.down()
 }
 
-exports.verify = function (next) {
+exports.verify = function () {
   assert.equal(db.numbers.indexOf(1), 0)
-  next()
 }